splitting functions and starting to manage columns
