+++ src/ft_print_data.c
@@ -6,13 +6,14 @@
 /*   By: irifarac <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/08 11:53:57 by irifarac          #+#    #+#             */
-/*   Updated: 2024/08/14 12:40:13 by irifarac         ###   ########.fr       */
+/*   Updated: 2024/08/15 21:59:30 by israel           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../inc/ft_ls.h"
 #include "ft_printf.h"
 #include "libft.h"
+#include "../inc/colors.h"
 
 typedef void	t_print(const char *name, const struct stat *stat, t_flags flag);
 static void	ft_iter(const char *dir_name, t_flags flags, bool print_dir, size_t count);
@@ -47,95 +48,13 @@ static void	ft_recursion(const char *dir_name, t_fileinfo *files, t_flags flags)
 	}
 }
 
-static void	ft_print_sbl(t_fileinfo *file)
-{
-	char	target[1024] = {0};
-	char	path[1024] = {0};
-	ssize_t	len;
-
-	len = 0;
-	ft_strcpy(path, file->fullpath);
-	if (ft_strcmp(path, "/") != 0)
-		ft_strcat(path, "/");
-	ft_strcat(path, file->name);
-	ft_printf(1, "%s -> ", file->name);
-	len = readlink(path, target, sizeof(target));
-	if (len < 0)
-	{
-		ft_printf(2, "readlink error: %s\n", strerror(errno));
-		return ;
-	}
-	target[len] = '\0';
-	ft_printf(1, "%s\n", target);
-}
-
-static void
-t_print_file(t_fileinfo *file, struct stat *statbuf, t_flags flags)
-{
-	struct passwd	*pwd;
-	struct group	*grp;
-	char		*formatted_time;
-
-	if (flags.long_format == true)
-	{
-		if (S_ISDIR(statbuf->st_mode))
-			ft_printf(1, "d");
-		else if (S_ISLNK(statbuf->st_mode))
-			ft_printf(1, "l");
-		else if (S_ISCHR(statbuf->st_mode))
-			ft_printf(1, "c");
-		else if (S_ISBLK(statbuf->st_mode))
-			ft_printf(1, "b");
-		else if (S_ISFIFO(statbuf->st_mode))
-			ft_printf(1, "p");
-		else if (S_ISSOCK(statbuf->st_mode))
-			ft_printf(1, "s");
-		else
-			ft_printf(1, "-");
-		ft_printf(1, (statbuf->st_mode & S_IRUSR) ? "r" : "-");
-		ft_printf(1, (statbuf->st_mode & S_IWUSR) ? "w" : "-");
-		ft_printf(1, (statbuf->st_mode & S_IXUSR) ? "x" : "-");
-		ft_printf(1, (statbuf->st_mode & S_IRGRP) ? "r" : "-");
-		ft_printf(1, (statbuf->st_mode & S_IWGRP) ? "w" : "-");
-		ft_printf(1, (statbuf->st_mode & S_IXGRP) ? "x" : "-");
-		ft_printf(1, (statbuf->st_mode & S_IROTH) ? "r" : "-");
-		ft_printf(1, (statbuf->st_mode & S_IWOTH) ? "w" : "-");
-		ft_printf(1, (statbuf->st_mode & S_IXOTH) ? "x" : "-");
-		ft_printf(1," %d ", statbuf->st_nlink);
-		if (flags.print_owner == true)
-		{
-			pwd = getpwuid(statbuf->st_uid);
-			ft_printf(1, "%s ", pwd->pw_name);
-		}
-		grp = getgrgid(statbuf->st_gid);
-		ft_printf(1 ,"%s ", grp->gr_name);
-		if (S_ISCHR(statbuf->st_mode) || S_ISBLK(statbuf->st_mode))
-			ft_printf(1, "%d, %d ", major(statbuf->st_rdev), minor(statbuf->st_rdev));
-		else
-			ft_printf(1, "%ld ", statbuf->st_size);
-		formatted_time = ctime(&statbuf->st_mtime);
-		formatted_time[16] = '\0';
-		ft_printf(1, "%s ", formatted_time + 4);
-		if (S_ISLNK(statbuf->st_mode))
-			ft_print_sbl(file);
-#ifdef BONUS
-		else
-			ft_printf(1, "%s\n", file->name);
-#else
-		else
-			ft_printf(1, "%s\n", file->name);
-#endif
-	}
-	else
-		ft_printf(1, "%s", file->name);
-}
 
 static void	ft_print_no_sort(struct stat statbuf, char *file, const char *dir_name, t_flags flags)
 {
 	t_fileinfo	*files;
 
 	files = ft_build_fileinfo(NULL, statbuf, file, dir_name);
-	t_print_file(files, &files->stat, flags);
+	ft_print_file(files, &files->stat, flags);
 	write(1, " ", 1);
 	free((char *)files->name);
 	free((char *)files->fullpath);
@@ -211,19 +130,13 @@ static void	ft_iter(const char *dir_name, t_flags flags, bool print_dir, size_t 
 
 void	ft_print_data(t_fileinfo *files, t_directory *dir, t_flags flags)
 {
-	t_fileinfo	*tmp;
 	t_directory	*tmp_dir;
 	size_t		count;
 	bool		print_dir;
+	size_t		max_len;
 
-	tmp = files;
-	while (tmp)
-	{
-		t_print_file(tmp, &tmp->stat, flags);
-		if (flags.long_format == false)
-			write(1, " ", 1);
-		tmp = tmp->next;
-	}
+	max_len = ft_max_len(files, dir);
+	ft_iter_file(files, flags, max_len);
 	if (flags.long_format == false && files)
 		write(1, "\n", 1);
 	tmp_dir = dir;
