+++ src/main.c
@@ -6,7 +6,7 @@
 /*   By: irifarac <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/24 11:29:26 by irifarac          #+#    #+#             */
-/*   Updated: 2024/08/14 11:10:04 by irifarac         ###   ########.fr       */
+/*   Updated: 2024/08/15 20:13:10 by israel           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,12 +19,14 @@ int	main(int argc, char **argv)
 	t_flags		flags;
 	t_fileinfo	*files;
 	t_directory	*dir;
+	struct winsize	size;
 
 	ft_memset(&files, 0, sizeof(t_fileinfo));
 	ft_memset(&dir, 0, sizeof(t_directory));
 	ft_memset(&flags, 0, sizeof(t_flags));
 	flags.print_owner = true;
 	ft_parse_flags(argc, argv, &flags);
+	ioctl(STDOUT_FILENO, TIOCGWINSZ, &size);
 	ft_parse(argc, argv, &files, &dir, flags);
 	if (files)
 		ft_sort_files(&files, flags);
